+++ src/components/FeedComponent.js
@@ -25,8 +25,8 @@ export default function FeedComponent({
   const navigation = useNavigation();
   const imagePost = data.image[0]
   const onPressFeed = () => {
-    // navigation.navigate(StackNav.PostComments,
-    //   {idPost: data});
+    navigation.navigate(StackNav.SinglePost,
+      {dataPost: data});
   }
 
   return (
