+++ src/navigation/Type/StackNavigation.js
@@ -214,6 +214,10 @@ export default function StackNavigation() {
         name={StackNav.EditPost}
         component={StackRoute.EditPost}
       />
+      <Stack.Screen
+        name={StackNav.SinglePost}
+        component={StackRoute.SinglePost}
+      />
     </Stack.Navigator>
   );
 }
